Extract document slicing into metadata plugin
